Moved Pages into components
Refactored MyMSU and Form Components in order to remove pages folder.
