+++ frontend/src/App.js
@@ -1,6 +1,5 @@
 import React, { Component } from "react";
 // TODO: Move files in pages to components.
-import Registration from "./pages/Registration";
 import { BrowserRouter as Router, Route } from "react-router-dom";
 import Wrapper from "./building_components/Wrapper/Wrapper";
 import { StoreProvider } from "./utils/GlobalState";
@@ -15,11 +14,9 @@ import DonatePage from "./components/Donate";
 import ShcolarshipPage from "./components/Scholarships";
 import AboutPage from "./components/About";
 import AdminPage from "./components/Admin";
-import AccountsPage from "./components/Accounts";
 import EditContactInfo from "./components/MyMSU/EditContactInfo";
 import EditMemberInfo from "./components/MyMSU/EditMemberInfo";
 import PayMembershipDues from "./components/MyMSU/PayMembershipDues.js";
-import Footer from "./building_components/Footer/Footer";
 import NAVBAR from "./building_components/Navbar/Navbar";
 import * as ROUTES from "./constants/routes";
 import { withAuthentication } from "./components/Session";
@@ -30,14 +27,12 @@ class App extends Component {
       <Router>
         <StoreProvider>
           <NAVBAR />
-          <Route exact path="/register" component={Registration} />
           <Route exact path="/" component={HomePage}></Route>
           <Route path={ROUTES.SIGN_UP} component={SignUpPage} />
           <Route path={ROUTES.SIGN_IN} component={SignInPage} />
           <Route path={ROUTES.PASSWORD_FORGET} component={PasswordForgetPage} />
           <Route path={ROUTES.HOME} component={HomePage} />
           <Route path={ROUTES.MYMSU} component={MyMsuPage} />
-          <Route path={ROUTES.ACCOUNTS} component={AccountsPage} />
           <Route path={ROUTES.EVENTS} component={EventsPage} />
           <Route path={ROUTES.ADMIN} component={AdminPage} />
           <Route path={ROUTES.DONATE} component={DonatePage} />
@@ -46,7 +41,6 @@ class App extends Component {
           <Route path={ROUTES.EDIT} component={EditContactInfo} />
           <Route path={ROUTES.EDIT_MEMBER_INFO} component={EditMemberInfo} />
           <Route path={ROUTES.PAY_DUES} component={PayMembershipDues} />
-          <Footer />
         </StoreProvider>
       </Router>
     );
