--- frontend/src/components/Accounts/index.js
@@ -1,36 +1,0 @@
-import React from 'react';
-// import { compose } from 'recompose';
-
-// import {
-//   AuthUserContext,
-//   withAuthorization,
-//   withEmailVerification,
-// } from '../Session';
-// import { withFirebase } from '../Firebase';
-// import { PasswordForgetForm } from '../PasswordForget';
-// import PasswordChangeForm from '../PasswordChange';
-
-// const SIGN_IN_METHODS = [
-//   {
-//     id: 'password',
-//     provider: null,
-//   },
-//   {
-//     id: 'google.com',
-//     provider: 'googleProvider',
-//   },
-//   {
-//     id: 'facebook.com',
-//     provider: 'facebookProvider',
-//   },
-//   {
-//     id: 'twitter.com',
-//     provider: 'twitterProvider',
-//   },
-// ];
-
-const AccountsPage = () => (
-<h1>Accounts</h1>
-);
-
-export default AccountsPage;
