+++ frontend/src/components/FormComponents/index.js
@@ -1,0 +1,13 @@
+import AddressForm from "./AddressForm";
+import ContactInfoForm from "./ContactInfoForm";
+import MemberInfoForm from "./MemberInfoForm";
+import MemberTypeForm from "./MemberTypeForm";
+
+export default ContactInfoForm;
+
+export {
+  AddressForm,
+  ContactInfoForm,
+  MemberInfoForm,
+  MemberTypeForm
+}
