+++ frontend/src/components/Home/index.js
@@ -1,6 +1,5 @@
 import React from "react";
 import Carousel from "../../building_components/Carousel/Carousel";
-import Footer from "../../building_components/Footer/Footer";
 
 const HomePage = () => (
   <div className="mainPage">
