+++ frontend/src/components/LoginManagement/index.js
@@ -1,0 +1,3 @@
+import LoginManagement from "./LoginManagement"
+
+export default LoginManagement;
