+++ frontend/src/components/MyMSU/MemberType.js
@@ -2,6 +2,7 @@ import React from "react";
 import Button from "../../building_components/Button";
 import { Link } from "react-router-dom";
 import * as ROUTES from "../../constants/routes";
+
 const MemberType = props => {
   console.log(props);
